Ensure that Munge service is running
It's easy to get into a situation where the Munge service is
enabled and configured but not actually running.  After the
handler that restarts the Munge service has been flushed,
ensure the Munge service is enabled and running.  This logic
ensures Munge is started (if it is not running) but does not
restart it if is running already.